requirements.txt
discord.py>=2.3.0
cerebras-cloud-sdk>=1.0.0
flask>=3.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
